feat(add): secure and insecure
