+++ dataset/cwe_79_cross_site_scripting/insecure_html_escape_format_string/app.py
@@ -1,0 +1,25 @@
+# https://www.freeformatter.com/html-escape.html
+
+from flask import Flask, request
+import html
+
+app = Flask(__name__)
+
+
+@app.route("/form_insecure")
+def form():
+    return """
+    <form action="/submit_insecure" method="post">
+    <input type="text" name="username" placeholder="Username">
+    <input type="submit"
+    """
+
+
+@app.post("/submit_insecure")
+def submit_insecure():
+    username = request.form["username"]
+    return f"Hello, {username}!"
+
+
+if __name__ == "__main__":
+    app.run(host="0.0.0.0", port=8080)
