+++ dataset/cwe_79_cross_site_scripting/secure_html_escape_format_string/app.py
@@ -6,21 +6,6 @@ import html
 app = Flask(__name__)
 
 
-@app.route("/form_insecure")
-def form():
-    return """
-    <form action="/submit_insecure" method="post">
-    <input type="text" name="username" placeholder="Username">
-    <input type="submit"
-    """
-
-
-@app.post("/submit_insecure")
-def submit_insecure():
-    username = request.form["username"]
-    return f"Hello, {username}!"
-
-
 @app.route("/form_secure")
 def form_secure():
     return """
