+++ dataset/cwe_79_cross_site_scripting/secure_html_escape_format_string/requirements.txt
@@ -1,0 +1,1 @@
+Flask==3.1.1
